config template and instanitation in init and sash; server-side include in httpd.

+++ Src/uClinux-dist/user/sash/cp.c
@@ -7,7 +7,13 @@ do_cp(argc, argv)
 	char	*srcname;
 	char	*destname;
 	char	*lastarg;
-
+	struct	stat	statbuf;
+	
+	BOOL     skipexisting = (argc > 1)  && (!strcmp(argv[1], "-se"));
+	if (skipexisting) {
+	    argc--;
+		argv++;
+	}
 	lastarg = argv[argc - 1];
 
 	dirflag = isadir(lastarg);
@@ -21,7 +27,8 @@ do_cp(argc, argv)
 		destname = lastarg;
 		if (dirflag)
 			destname = buildname(destname, srcname);
-
-		(void) copyfile(*++argv, destname, FALSE);
+			
+		if ( (!skipexisting) || (stat(destname, &statbuf) < 0) )
+			(void) copyfile(*++argv, destname, FALSE);
 	}
 }
